Initial version of BS loader. refs #13535
